requirement.txt
langchain
langchain-openai
langchain-ollama 
openai 
python-dotenv 